move preview panel to common folder
